(SIMP-5754) Fix `pup_4` Gitlab CI cache poisoning (#45)
This patch fixes a `.gitlab-ci.yml` bug that was poisoning the GitLab
Runners' RubyGems cache.  The `pup4` YAML anchor must use `image:
'ruby:2.1'` instead of `image: 'ruby:2.4'`.

Note: This commit is part of a common asset update, and adjusts other
aspects of the `.gitlab-ci.yml` file in order to bring it in line with a
standardized pipeline.  In particular, PE 2017.3 was recently
deprecated, removing Puppet 5.3 from the test matrix of most modules.

This patch also updates the version for the complete fix to the CHANGELOG
 EOL problem. The last tagged version has DOS EOLs in the
CHANGELOG. 

SIMP-6016 #close
SIMP-5754 #comment updated pupmod-simp-postfix